requirements.txt
pandas
numpy 
tqdm
torch
torch-geometric
networkx
PyYAML
scikit_learn
matplotlib
scipy